TS: IDA/options + removing idaLabel from redux state
